Only run ldconfig when codecs relocated
ldconfig can be very slow, so it should only be run if new codecs were
relocated. The rsync -i|--itemize-changes option can be used to see if
there were any changes when syncing the directories. If the output is
empty, there were not any changes and ldconfig can be skipped.

This also adds -X to the ldconfig call so that it doesn't relink the .so
symlinks and only rebuilds the cache. This slows things down
unnecessarily since we already manage the symlinks for the codecs.

[endlessm/eos-shell#5503]