Only relocate codecs on updates
Since the codecs to relocate are shipped with ostree, it makes no sense
to try relocation any time other than after an update.

Use systemd's ConditionNeedsUpdate=/var to detect when an update is
needed. In order to run before systemd-update-done.service, this needs
to be moved to sysinit.target with dependencies explicitly listed.

[endlessm/eos-shell#5502]