Make eos-relocate-codecs handle encrypted codecs packs under /usr
From now on, codecs will be shipped as encrypted 'packs' that will be
located under /usr/share/EndlessOS/codecs, instead of plain files. So,
we need to decrypt those codecs (if the relevant GPG key is available),
unpack their contents and place then in their right location.

To handle this, we will use a new directory /var/lib/codecs/.gnupg as
the GNUPGHOME for importing keys and decrypting, as well a drop-in
subdirectory /var/lib/codecs/.gnupg/keys.d, where the relevant private
keys will be searched for, so that they can be imported and used.

[endlessm/eos-shell#5767]