Use a status file to avoid wasting unnecessary time during boot
With the new GPG-based model implemented, we need to do quite a few
things before calling rsync (decrypt + uncompress), that would always
artificially increase boot time for no good reason.

To prevent that situation from happening, simply store a small status
file with the list of files under /usr/share/EndlessOS/codecs and their
respective MD5 checkums, so that we can compare the current status with
the data stored on that file on every boot, early returning if possible.

[endlessm/eos-shell#5767]