Handle pre-installed GPG keys under /usr/share/Endless/codecs
In order to get users of EOS <= 2.4 updated to the new mechanism to
distribute codecs, we will be temporarily shipping the GPG keys needed
to decrypt them under /usr/share/Endless/codecs, so we need to have
this script to look for them and copy them to their right place.

Note that this changeset will only be valid during the life cycle of
the EOS 2.5 series, and should be removed afterwards.

[endlessm/eos-shell#5812]