Use the available checksums to check correctness for deployed files
From now on, codecs pack CAN ship an additional .metadata subdirectory
together with the actual files, where additional information can be
provided, such as checksums for the codecs files, for instance, so that
we can validate their correctness on runtime.

Note that for now we are only expecting a *.md5sum file in there, but
I created a .metadata directory for it anyway, just in case. Also, this
is an OPTIONAL feature: if the *md5sum file is not present it is simply
ignored, and no additional tasks are performed.

[endlessm/eos-shell#6090]