Write a helper binary for GStreamer codecs installer, to report metrics
This will be called whenever Totem or Rhythmbox tries to play a media file
we don't have the right codec available for, giving us the opportunity to
do whatever we need to do at that point.

At the moment, we will just report the "missing codec" event via EOSMetrics,
but in the future we might want to do other things, such as actually support
the installation of the missing codec in EOS, or hinting the user into maybe
installing a 3rd-party application from the App Center that can open the file.

https://phabricator.endlessm.com/T6405