correct error from printing freed variant
Now we're using the _sync method call the GVariant's floating ref is sunk by the event recorder, so payload is free by the time the method returns. This prints it to a string before calling the event recorder so we can still print the debug message *and* send the metric.